+++ packages/article-magazine-standard/__tests__/web/semantic.web.test.js
@@ -11,11 +11,13 @@ import {
   replacePropTransform,
   replaceTransform
 } from "@times-components/jest-serializer";
-import Context from "@times-components/context";
+import { ContextProviderWithDefaults } from "@times-components/context";
 import { scales } from "@times-components/styleguide";
 import Article from "../../src/article-magazine-standard";
 import articleFixture, { testFixture } from "../../fixtures/full-article";
 import { adConfig } from "../ad-mock";
+
+jest.mock("react-helmet-async", () => ({ Helmet: "Helmet" }));
 
 const omitProps = new Set([
   "className",
@@ -126,9 +128,8 @@ const tests = [
       const scale = scales.large;
       const sectionColour = "#FFFFFF";
       const testRenderer = TestRenderer.create(
-        <Context.Provider
+        <ContextProviderWithDefaults
           value={{
-            makeArticleUrl: () => "https://some-url.io",
             theme: { scale, sectionColour }
           }}
         >
@@ -146,7 +147,7 @@ const tests = [
             receiveChildList={() => {}}
             spotAccountId=""
           />
-        </Context.Provider>
+        </ContextProviderWithDefaults>
       );
 
       expect(testRenderer).toMatchSnapshot();