+++ packages/context/src/context.js
@@ -1,6 +1,12 @@
-import { createContext } from "react";
+import React, { createContext } from "react";
 import defaults from "./defaults";
 import SectionContext from "./section-context";
 
-export default createContext(defaults);
-export { defaults, SectionContext };
+const Context = createContext(defaults);
+
+function ContextProviderWithDefaults({ value, ...props }) {
+  return <Context.Provider {...props} value={{ ...defaults, ...value }} />;
+}
+
+export default Context;
+export { defaults, SectionContext, ContextProviderWithDefaults };