+++ packages/provider-queries/src/article.web.js
@@ -14,6 +14,7 @@ export default addTypenameToDocument(gql`
       }
       commentsEnabled
       content: paywalledContent
+      descriptionMarkup: summary(maxCharCount: 200)
       dropcapsDisabled
       expirableFlags {
         type
@@ -65,6 +66,13 @@ export default addTypenameToDocument(gql`
         }
       }
       standfirst
+      tiles {
+        slices {
+          sections {
+            title
+          }
+        }
+      }
       template
       textColour {
         rgba {