+++ packages/ssr/src/lib/make-url.js
@@ -1,1 +1,0 @@
-module.exports = ({ id }) => `http://localhost:3000/article/${id}`;