+++ packages/ad/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.31](https://github.com/newsuk/times-components/compare/@times-components/ad@2.5.30...@times-components/ad@2.5.31) (2019-04-29)
+
+**Note:** Version bump only for package @times-components/ad
+
+
+
+
+
 ## [2.5.30](https://github.com/newsuk/times-components/compare/@times-components/ad@2.5.29...@times-components/ad@2.5.30) (2019-04-18)
 
 **Note:** Version bump only for package @times-components/ad