+++ packages/article-comments/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.8](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.5.7...@times-components/article-comments@0.5.8) (2019-04-29)
+
+**Note:** Version bump only for package @times-components/article-comments
+
+
+
+
+
 ## [0.5.7](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.5.6...@times-components/article-comments@0.5.7) (2019-04-24)
 
 **Note:** Version bump only for package @times-components/article-comments