+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.53](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.2.52...@times-components/error-view@2.2.53) (2019-04-29)
+
+**Note:** Version bump only for package @times-components/error-view
+
+
+
+
+
 ## [2.2.52](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.2.51...@times-components/error-view@2.2.52) (2019-04-18)
 
 **Note:** Version bump only for package @times-components/error-view