+++ packages/slice-layout/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.27.11](https://github.com/newsuk/times-components/compare/@times-components/slice-layout@0.27.10...@times-components/slice-layout@0.27.11) (2019-04-29)
+
+**Note:** Version bump only for package @times-components/slice-layout
+
+
+
+
+
 ## [0.27.10](https://github.com/newsuk/times-components/compare/@times-components/slice-layout@0.27.9...@times-components/slice-layout@0.27.10) (2019-04-18)
 
 **Note:** Version bump only for package @times-components/slice-layout